Make vertex type friendlier for graphics

+++ experiments/2024-10-30/src/mesh.rs
@@ -1,4 +1,4 @@
 pub struct Mesh {
-    pub vertices: Vec<[f64; 3]>,
+    pub vertices: Vec<[f32; 3]>,
     pub triangles: Vec<[u32; 3]>,
 }
